rename files sperl_yacc.? -> sperl_yacc_util.?, sperl_yacc.tab.? -> sperl_yacc.?.
* rename sperl_yacc.{c,h} -> sperl_yacc_util.{c,h}
* rename sperl_yacc.tab.{c,h} -> sperl_yacc.{c,h}
* Makefile: remove a suffix rule hack

After this commit, `make clean` will be needed.
