+++ .gitignore
@@ -1,8 +1,11 @@
 /sperl
 /libsperl.a
-*.tab.c
-*.tab.h
 core.*
 *.output
 /tmp_*
 /objs
+
+# generated by bison
+
+/sperl_yacc.h
+/sperl_yacc.c
