+++ sperl_toke.c
@@ -6,8 +6,8 @@
 #include "sperl.h"
 #include "sperl_parser.h"
 #include "sperl_toke.h"
+#include "sperl_yacc_util.h"
 #include "sperl_yacc.h"
-#include "sperl_yacc.tab.h"
 #include "sperl_op.h"
 #include "sperl_allocator.h"
 #include "sperl_constant.h"
